corrections to routes

+++ routes/api/product-routes.js
@@ -17,21 +17,33 @@ router.get('/', async (req, res) => {
 // GET one product
 router.get('/:id', async (req, res) => {
   // find a single product by its `id`
-  try {
-    const productData = await Product.findByPk(req.params.id, {
-      // be sure to include its associated Category and Tag data
-      include: [{ model: Category, through: ProductTag, as: 'tag_id'}]
-    });
+  // try {
+  //   const productData = await Product.findOne(req.params.id, {
+  //     // be sure to include its associated Category and Tag data
+  //     include: [{ model: Category, through: ProductTag, as: 'tag_id'}]
+  //   });
 
-    if (!productData) {
-      res.status(404).json({ message: 'No product found!' });
-      return;
-    }
+  //   if (!productData) {
+  //     res.status(404).json({ message: 'No product found!' });
+  //     return;
+  //   }
 
-    res.status(200).json(productData);
-  } catch (err) {
+  Product.findOne({
+    where: {
+      id: req.params.id
+    }, 
+      include: [
+        Category, 
+        {
+          model: Tag,
+          through: ProductTag
+        }
+      ]
+   }).then(products => res.json(products))
+  
+  .catch ((err) => {
     res.status(500).json(err);
-  }
+  })
 });
 
 // CREATE new product
@@ -43,12 +55,11 @@ router.post('/', async (req, res) => {
       stock: 3,
       tagIds: [1, 2, 3, 4]
     }
-    
   */
   Product.create(req.body)
     .then((product) => {
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-      if (req.body.tagIds.length) {
+      if (req.body.tagIds && req.body.tagIds.length) {
         const productTagIdArr = req.body.tagIds.map((tag_id) => {
           return {
             product_id: product.id,
@@ -72,38 +83,39 @@ router.put('/:id', (req, res) => {
   // update product data
   Product.update(req.body, {
     where: {
-      id: req.params.id,
-    },
+      id: req.params.id
+    }
   })
-    .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
+    .then(product => {
+      if (req.body.tagIds && req.body.tagIds.length) {
+        const productTags = ProductTag.findAll({
+          where: { product_id: req.params.id }
         });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        // create filtered list of new tag_ids
+        const newProductTags = req.body.tagIds
+          .filter(tag_id => !productTagIds.includes(tag_id))
+          .map(tag_id => {
+            return {
+              product_id: req.params.id,
+              tag_id
+            };
+          });
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags)
+        ]);
+      }
+
+      return res.json(product);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
+    .catch(err => {
       // console.log(err);
       res.status(400).json(err);
     });
